requirements.txt
asgiref
Django
django-cors-headers
django-debug-toolbar
django-filter

djangorestframework
djangorestframework-simplejwt
drf-flex-fields
importlib-metadata
Markdown==3.3

PyJWT

pytz
six
sqlparse
typing
zipp
